backend: pipe game and extractor logs to a file

+++ src-tauri/src/util/file.rs
@@ -1,12 +1,8 @@
 use std::path::PathBuf;
 
-pub fn delete_dir_or_folder(path: &PathBuf) -> Result<(), std::io::Error> {
-  if path.exists() {
-    if path.is_dir() {
-      std::fs::remove_dir_all(path)?;
-    } else {
-      std::fs::remove_file(path)?;
-    }
+pub fn delete_dir(path: &PathBuf) -> Result<(), std::io::Error> {
+  if path.exists() && path.is_dir() {
+    std::fs::remove_dir_all(path)?;
   }
   Ok(())
 }
@@ -16,5 +12,12 @@ pub fn create_dir(path: &PathBuf) -> Result<(), std::io::Error> {
     return Ok(());
   }
   std::fs::create_dir_all(path)?;
+  Ok(())
+}
+
+pub fn delete_file(path: &PathBuf) -> Result<(), std::io::Error> {
+  if path.exists() && path.is_file() {
+    std::fs::remove_file(path)?;
+  }
   Ok(())
 }
